+++ src/full/Agda/Utils/Semigroup.hs
@@ -8,9 +8,14 @@ import Data.Semigroup ( Semigroup, (<>) )
 
 import Control.Applicative (liftA2)
 
+import Control.Monad.Except (ExceptT)
 import Control.Monad.Reader (ReaderT)
 import Control.Monad.State  (StateT)
 import Control.Monad.Writer (WriterT)
+
+instance (Monad m, Semigroup doc)       => Semigroup (ExceptT e m doc) where
+  {-# INLINE (<>) #-}
+  (<>) = liftA2 (<>)
 
 instance (Applicative m, Semigroup doc) => Semigroup (ReaderT s m doc) where
   {-# INLINE (<>) #-}