+++ CHANGELOG.md
@@ -12,8 +12,8 @@ Installation
 Pragmas and options
 -------------------
 
-* New warning `RecursiveDisplayForm` instead of hard error
-  when a display form is recursive and thus illegal.
+* New warning `InvalidDisplayForm` instead of hard error
+  when a display form is illegal (and thus ignored).
 
 * New warning `WithClauseProjectionFixityMismatch` instead of hard error
   when in a with-clause a projection is used in a different fixity